// services/http-auth/src/generated/grpc/typescript/threads_pb.d.ts
    export type AsObject = {
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        timestamp?: google_protobuf_wrappers_pb.Int64Value.AsObject,
    }
// ...
export class ThreadModel extends jspb.Message { 

    hasId(): boolean;
    clearId(): void;
    getId(): google_protobuf_wrappers_pb.StringValue | undefined;
    setId(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasName(): boolean;
    clearName(): void;
    getName(): google_protobuf_wrappers_pb.StringValue | undefined;
    setName(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasDescription(): boolean;
    clearDescription(): void;
    getDescription(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDescription(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasContent(): boolean;
    clearContent(): void;
    getContent(): google_protobuf_wrappers_pb.StringValue | undefined;
    setContent(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;
    clearImagesList(): void;
    getImagesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setImagesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addImages(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;

    hasCreatedAt(): boolean;
    clearCreatedAt(): void;
    getCreatedAt(): google_protobuf_wrappers_pb.StringValue | undefined;
    setCreatedAt(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasUpdatedAt(): boolean;
    clearUpdatedAt(): void;
    getUpdatedAt(): google_protobuf_wrappers_pb.StringValue | undefined;
    setUpdatedAt(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasOwner(): boolean;
    clearOwner(): void;
    getOwner(): google_protobuf_wrappers_pb.StringValue | undefined;
    setOwner(value?: google_protobuf_wrappers_pb.StringValue): ThreadModel;

    hasDeleted(): boolean;
    clearDeleted(): void;
    getDeleted(): google_protobuf_wrappers_pb.BoolValue | undefined;
    setDeleted(value?: google_protobuf_wrappers_pb.BoolValue): ThreadModel;
    clearPromotedList(): void;
    getPromotedList(): Array<PromotedObject>;
    setPromotedList(value: Array<PromotedObject>): ThreadModel;
    addPromoted(value?: PromotedObject, index?: number): PromotedObject;
    clearDonationsList(): void;
    getDonationsList(): Array<DonationObject>;
    setDonationsList(value: Array<DonationObject>): ThreadModel;
    addDonations(value?: DonationObject, index?: number): DonationObject;
    clearLikesList(): void;
    getLikesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setLikesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addLikes(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;
    clearDislikesList(): void;
    getDislikesList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setDislikesList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addDislikes(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;
    clearTagsList(): void;
    getTagsList(): Array<google_protobuf_wrappers_pb.StringValue>;
    setTagsList(value: Array<google_protobuf_wrappers_pb.StringValue>): ThreadModel;
    addTags(value?: google_protobuf_wrappers_pb.StringValue, index?: number): google_protobuf_wrappers_pb.StringValue;

    hasLikesCount(): boolean;
    clearLikesCount(): void;
    getLikesCount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setLikesCount(value?: google_protobuf_wrappers_pb.UInt64Value): ThreadModel;

    hasDislikesCount(): boolean;
    clearDislikesCount(): void;
    getDislikesCount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setDislikesCount(value?: google_protobuf_wrappers_pb.UInt64Value): ThreadModel;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): ThreadModel.AsObject;
    static toObject(includeInstance: boolean, msg: ThreadModel): ThreadModel.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: ThreadModel, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): ThreadModel;
    static deserializeBinaryFromReader(message: ThreadModel, reader: jspb.BinaryReader): ThreadModel;
}

export namespace ThreadModel {
    export type AsObject = {
        id?: google_protobuf_wrappers_pb.StringValue.AsObject,
        name?: google_protobuf_wrappers_pb.StringValue.AsObject,
        description?: google_protobuf_wrappers_pb.StringValue.AsObject,
        content?: google_protobuf_wrappers_pb.StringValue.AsObject,
        imagesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        createdAt?: google_protobuf_wrappers_pb.StringValue.AsObject,
        updatedAt?: google_protobuf_wrappers_pb.StringValue.AsObject,
        owner?: google_protobuf_wrappers_pb.StringValue.AsObject,
        deleted?: google_protobuf_wrappers_pb.BoolValue.AsObject,
        promotedList: Array<PromotedObject.AsObject>,
        donationsList: Array<DonationObject.AsObject>,
        likesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        dislikesList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        tagsList: Array<google_protobuf_wrappers_pb.StringValue.AsObject>,
        likesCount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
        dislikesCount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
    }
}

export class PromotedObject extends jspb.Message { 

    hasPromoter(): boolean;
    clearPromoter(): void;
    getPromoter(): google_protobuf_wrappers_pb.StringValue | undefined;
    setPromoter(value?: google_protobuf_wrappers_pb.StringValue): PromotedObject;

    hasDate(): boolean;
    clearDate(): void;
    getDate(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDate(value?: google_protobuf_wrappers_pb.StringValue): PromotedObject;

    hasAmount(): boolean;
    clearAmount(): void;
    getAmount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setAmount(value?: google_protobuf_wrappers_pb.UInt64Value): PromotedObject;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): PromotedObject.AsObject;
    static toObject(includeInstance: boolean, msg: PromotedObject): PromotedObject.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: PromotedObject, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): PromotedObject;
    static deserializeBinaryFromReader(message: PromotedObject, reader: jspb.BinaryReader): PromotedObject;
}

export namespace PromotedObject {
    export type AsObject = {
        promoter?: google_protobuf_wrappers_pb.StringValue.AsObject,
        date?: google_protobuf_wrappers_pb.StringValue.AsObject,
        amount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
    }
}

export class DonationObject extends jspb.Message { 

    hasDonator(): boolean;
    clearDonator(): void;
    getDonator(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDonator(value?: google_protobuf_wrappers_pb.StringValue): DonationObject;

    hasDate(): boolean;
    clearDate(): void;
    getDate(): google_protobuf_wrappers_pb.StringValue | undefined;
    setDate(value?: google_protobuf_wrappers_pb.StringValue): DonationObject;

    hasAmount(): boolean;
    clearAmount(): void;
    getAmount(): google_protobuf_wrappers_pb.UInt64Value | undefined;
    setAmount(value?: google_protobuf_wrappers_pb.UInt64Value): DonationObject;

    serializeBinary(): Uint8Array;
    toObject(includeInstance?: boolean): DonationObject.AsObject;
    static toObject(includeInstance: boolean, msg: DonationObject): DonationObject.AsObject;
    static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
    static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
    static serializeBinaryToWriter(message: DonationObject, writer: jspb.BinaryWriter): void;
    static deserializeBinary(bytes: Uint8Array): DonationObject;
    static deserializeBinaryFromReader(message: DonationObject, reader: jspb.BinaryReader): DonationObject;
}

export namespace DonationObject {
    export type AsObject = {
        donator?: google_protobuf_wrappers_pb.StringValue.AsObject,
        date?: google_protobuf_wrappers_pb.StringValue.AsObject,
        amount?: google_protobuf_wrappers_pb.UInt64Value.AsObject,
    }
}